# perf: avoid allocs in unicodeFoldTransformer