@lithammer 2023-05-03
Released as part of [v1.1.7](https://github.com/lithammer/fuzzysearch/releases/tag/v1.1.7)